backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
alembic>=1.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
httpx>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.12
aiosqlite>=0.20.0
greenlet>=3.0.0
python-docx>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
